pyproject.toml
[project]
name = "corporate-meeting-minutes"
version = "0.1.0"
description = ""
authors = [
    { name = "Derek Pappas", email = "derekepappas@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.2.0,<2.0.0",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "pyobjc-core (>=12.1,<13.0)",
    "pyobjc-framework-vision (>=12.1,<13.0)",
    "pyobjc-framework-quartz (>=12.1,<13.0)",
    "pyobjc-framework-cocoa (>=12.1,<13.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

